+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **ToDo**
 - improve/replace in-game graphics
 - maybe increase speed and/or difficulty
+- maybe save highscore on server-side
 
 **Changes**
 - added health indicator
@@ -14,3 +15,5 @@
 - added multiple enemy types (e.g. alien ships and asteroids)
 - added health to enemies and asteroids (e.g. destroy objects only if hit by multiple shots)
 - added option to toggle fullscreen mode
+- added logo
+- save current highscore as cookie